requirements.txt
django>=4.2.0
openai-whisper
torch
torchaudio
numpy
python-dotenv
pillow
